# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Inventory API
  version: 1.0.0
tags:
  - name: Items
    description: Operations for managing inventory items
  - name: Orders
    description: Operations for managing orders
paths:
  /api/items:
    get:
      tags:
        - Items
      summary: Get all items
      operationId: getAllItems
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      tags:
        - Items
      summary: Create a new item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '201':
          description: The created item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /api/items/{id}:
    get:
      tags:
        - Items
      summary: Get an item by ID
      operationId: getItemById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    put:
      tags:
        - Items
      summary: Update an item
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: The updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      tags:
        - Items
      summary: Delete an item by ID
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
  /api/orders:
    get:
      tags:
        - Orders
      summary: Get all orders
      operationId: getAllOrders
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags:
        - Orders
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: The created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get an order by ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    put:
      tags:
        - Orders
      summary: Update an existing order
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '200':
          description: The updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/{id}/purchase:
    post:
      tags:
        - Orders
      summary: Purchase an existing order
      operationId: purchaseOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: virtual
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: The purchased order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        quantity:
          type: integer
    OrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        deliveryAddress:
          type: string
    OrderItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        quantity:
          type: integer
    OrderStatus:
      type: string
      enum:
        - SAVED
        - PURCHASED
        - HELD
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        deliveryAddress:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'